Plugin - MSS Controller - Work towards rework of things
