+++ software/plugins/mss-controller/build.gradle
@@ -25,7 +25,7 @@ dependencies {
     implementation 'io.quarkus:quarkus-micrometer'
     implementation 'io.quarkus:quarkus-smallrye-openapi'
 
-    implementation 'tech.ebp.oqm.lib:core-api-lib-quarkus:1.0.0-SNAPSHOT'
+    implementation 'tech.ebp.oqm.lib:core-api-lib-quarkus:2.1.0-SNAPSHOT'
 
     implementation 'com.fazecast:jSerialComm:2.10.1'
     implementation 'com.github.docker-java:docker-java:3.3.3'
