+++ software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/interfaces/rest/VoiceInteraction.java
@@ -1,20 +1,15 @@
 package com.ebp.openQuarterMaster.plugin.interfaces.rest;
 
-import com.ebp.openQuarterMaster.plugin.moduleInteraction.ItemVoiceSearchResults;
-import com.ebp.openQuarterMaster.plugin.moduleInteraction.command.response.CommandResponse;
 import com.ebp.openQuarterMaster.plugin.moduleInteraction.service.VoiceSearchService;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.enterprise.context.RequestScoped;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
-import jakarta.ws.rs.PathParam;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.http.entity.ContentType;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.openapi.annotations.tags.Tags;
