+++ software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/model/module/command/CommandType.java
@@ -1,4 +1,4 @@
-package com.ebp.openQuarterMaster.plugin.moduleInteraction.command;
+package com.ebp.openQuarterMaster.plugin.model.module.command;
 
 public enum CommandType {
 	GET_MODULE_INFO,
