+++ software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/model/module/command/GetModInfoCommand.java
@@ -1,11 +1,11 @@
-package com.ebp.openQuarterMaster.plugin.moduleInteraction.command;
+package com.ebp.openQuarterMaster.plugin.model.module.command;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class GetModInfoCommand extends MssCommand{
+public class GetModInfoCommand extends MssCommand {
 	@Getter
 	private static final GetModInfoCommand instance = new GetModInfoCommand();
