+++ software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/model/module/command/MssCommand.java
@@ -1,10 +1,9 @@
-package com.ebp.openQuarterMaster.plugin.moduleInteraction.command;
+package com.ebp.openQuarterMaster.plugin.model.module.command;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 @AllArgsConstructor
