+++ software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/model/module/command/response/CommandResponseStatus.java
@@ -1,0 +1,7 @@
+package com.ebp.openQuarterMaster.plugin.model.module.command.response;
+
+public enum CommandResponseStatus {
+	OK,
+	R_ERR,
+	ERR
+}
