+++ software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/model/module/command/response/GetModuleInfoResponse.java
@@ -1,4 +1,4 @@
-package com.ebp.openQuarterMaster.plugin.moduleInteraction.command.response;
+package com.ebp.openQuarterMaster.plugin.model.module.command.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
