--- software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/moduleInteraction/command/response/CommandResponseStatus.java
@@ -1,7 +1,0 @@
-package com.ebp.openQuarterMaster.plugin.moduleInteraction.command.response;
-
-public enum CommandResponseStatus {
-	OK,
-	R_ERR,
-	ERR
-}
