+++ software/plugins/mss-controller/src/main/java/com/ebp/openQuarterMaster/plugin/moduleInteraction/module/serialModule/exceptions/CommandReturnedErrorException.java
